Move the sidecar into its own package
And replace the environment variables with flags.

This clears the way to integrate the operator into the code base.
