+++ README.md
@@ -5,6 +5,7 @@ cloud provider from [Vault](https://www.vaultproject.io) and serves them via
 http to be consumed by the cloud-provider's SDK.
 
 It is intended to be used with a Vault setup like [this](https://github.com/utilitywarehouse/vault-manifests).
+
 The sidecar logs with Vault using its Kubernetes Service Account, requests
 credentials from a secrets engine and serves the acquired credentials via http.
 The sidecar detects the lease expiration and keeps the served credentials
@@ -19,33 +20,54 @@ Supported providers (secret engines):
 - `aws`
 - `gcp`
 
-## Environment variables
+For `aws`:
 
-Configuration is provided to the sidecar via the environment.
+```
+./vault-kube-cloud-credentials \
+  -aws-sidecar \
+  -kube-auth-role=<kubernetes auth role> \
+  -aws-role=<aws secret role>
+```
 
-Required:
+And `gcp`:
 
-- `VKAC_KUBE_AUTH_ROLE`: kubernetes auth backend role used to login
+```
+./vault-kube-cloud-credentials \
+  -gcp-sidecar \
+  -kube-auth-role=<kubernetes auth role> \
+  -gcp-roleset=<gcp secret roleset>
+```
 
-and one of the following:
+Refer to the usage for more options:
 
-- `VKAC_AWS_SECRET_ROLE`: aws secret backend role to retrieve credentials with
-- `VKAC_GCP_SECRET_ROLESET`: gcp secret backend roleset to retrieve credentials with
-
-These variables determine which secret engine is used and the format of the
-credentials which are served over http.
-
-Optional:
-
-- `VKAC_AWS_SECRET_BACKEND_PATH`: path of the aws secret backend (default: `aws`)
-- `VKAC_AWS_SECRET_ROLE_ARN`: the ARN of the role to assume. This is required
-  when there is more than one `role_arn` configured against the backend role
-  (not set by default)
-- `VKAC_GCP_SECRET_BACKEND_PATH`: path of the gcp secret backend (default: `gcp`)
-- `VKAC_KUBE_AUTH_BACKEND_PATH`: path of the kubernetes auth backend (default: `kubernetes`)
-- `VKAC_KUBE_SA_TOKEN_PATH`: path to a file containing the Kubernetes service account token (default: `/var/run/secrets/kubernetes.io/serviceaccount/token`)
-- `VKAC_LISTEN_HOST`: host to bind to (default: `127.0.0.1`)
-- `VKAC_LISTEN_PORT`: port to bind to (default: `8000`)
+```
+./vault-kube-cloud-credentials -h
+Usage of ./vault-kube-cloud-credentials:
+  -aws-role string
+    	AWS secret role (required when -aws-sidecar is set)
+  -aws-role-arn string
+    	AWS role arn to assume
+  -aws-secret-backend string
+    	AWS secret backend path (default "aws")
+  -aws-sidecar
+    	Run the AWS sidecar
+  -gcp-roleset string
+    	GCP roleset (required when -gcp-sidecar is set)
+  -gcp-secret-backend string
+    	GCP secret backend path (default "gcp")
+  -gcp-sidecar
+    	Run the GCP sidecar
+  -kube-auth-backend string
+    	Kubernetes auth backend path (default "kubernetes")
+  -kube-auth-role string
+    	Kubernetes auth role (required when -aws-sidecar or -gcp-sidecar are set)
+  -kube-token-path string
+    	Path to the kubernetes serviceaccount token (default "/var/run/secrets/kubernetes.io/serviceaccount/token")
+  -listen-host string
+    	Host to listen on (default "127.0.0.1")
+  -listen-port string
+    	Port to listen on (default "8000")
+```
 
 Additionally, you can use any of the [environment variables supported by the Vault
 client](https://www.vaultproject.io/docs/commands/#environment-variables), most
